+++ script/build.ps1
@@ -13,7 +13,7 @@ $GIT_SHA = git rev-parse --short HEAD
 $VERSION = git rev-parse --abbrev-ref HEAD
 $LDFLAGS = "-X github.com/dremio/dremio-diagnostic-collector/pkg/versions.GitSha=$GIT_SHA -X github.com/dremio/dremio-diagnostic-collector/pkg/versions.Version=$VERSION"
 
-New-Item -ItemType File -Path ./cmd/root/ddcbinary/output/ddc.zip -Force
+New-Item -ItemType File -Path .\cmd\root\ddcbinary\output\ddc.zip -Force
 # This assumes that you have 'go' installed in your environment
 $env:GOOS="linux"
 $env:GOARCH="amd64"
@@ -21,7 +21,7 @@ go build -ldflags "$LDFLAGS" -o .\bin\ddc
 
 # Use Compress-Archive to create zip file and then move it
 Compress-Archive -Path .\bin\ddc -DestinationPath .\bin\ddc.zip
-Move-Item -Path .\bin\ddc.zip -Destination .\cmd\root\ddcbinary\output\ddc.zip
+Move-Item -Force -Path  .\bin\ddc.zip -Destination .\cmd\root\ddcbinary\output\ddc.zip
 Remove-Item -Path .\bin\ddc
 
 $env:GOOS="windows"