+++ script/release-build.ps1
@@ -11,11 +11,11 @@ $GIT_SHA = git rev-parse --short HEAD
 $VERSION = $args[0]
 $LDFLAGS = "-X github.com/dremio/dremio-diagnostic-collector/pkg/versions.GitSha=$GIT_SHA -X github.com/dremio/dremio-diagnostic-collector/pkg/versions.Version=$VERSION"
 
-Write-Output "Cleaning bin folder…"
+Write-Output "Cleaning bin folder"
 Get-Date -Format "HH:mm:ss"
 .\script\clean
 
-Write-Output "Building embedded binary linux-amd64…"
+Write-Output "Building embedded binary linux-amd64"
 Get-Date -Format "HH:mm:ss"
 
 $env:GOOS="linux"
@@ -25,36 +25,42 @@ go build -ldflags "$LDFLAGS" -o ./bin/ddc
 Copy-Item -Path ./default-ddc.yaml -Destination ./bin/ddc.yaml
 Compress-Archive -Path ./bin/ddc -DestinationPath ./bin/ddc.zip
 Remove-Item ./bin/ddc
-Move-Item -Path ./bin/ddc.zip -Destination ./cmd/root/ddcbinary/output/ddc.zip
-Move-Item -Path ./bin/ddc.yaml -Destination ./bin/ddc.yaml
+Move-Item -Force -Path ./bin/ddc.zip -Destination ./cmd/root/ddcbinary/output/ddc.zip
 
-Write-Output "Building linux-amd64…"
+Write-Output "Building linux-amd64"
 Get-Date -Format "HH:mm:ss"
 go build -ldflags "$LDFLAGS" -o ./bin/ddc
 Compress-Archive -Path ./bin/ddc, ./bin/ddc.yaml -DestinationPath ./bin/ddc-linux-amd64.zip
 
-Write-Output "Building linux-arm64…"
+Write-Output "Building linux-arm64"
 Get-Date -Format "HH:mm:ss"
 $env:GOARCH="arm64"
 go build -ldflags "$LDFLAGS" -o ./bin/ddc
 Compress-Archive -Path ./bin/ddc, ./bin/ddc.yaml -DestinationPath ./bin/ddc-linux-arm64.zip
 
-Write-Output "Building darwin-os-x-amd64…"
+Write-Output "Building darwin-os-x-amd64"
 Get-Date -Format "HH:mm:ss"
 $env:GOOS="darwin"
 $env:GOARCH="amd64"
 go build -ldflags "$LDFLAGS" -o ./bin/ddc
 Compress-Archive -Path ./bin/ddc, ./bin/ddc.yaml -DestinationPath ./bin/ddc-mac-intel.zip
 
-Write-Output "Building darwin-os-x-arm64…"
+Write-Output "Building darwin-os-x-arm64"
 Get-Date -Format "HH:mm:ss"
 $env:GOARCH="arm64"
 go build -ldflags "$LDFLAGS" -o ./bin/ddc
 Compress-Archive -Path ./bin/ddc, ./bin/ddc.yaml -DestinationPath ./bin/ddc-mac-m-series.zip
 
-Write-Output "Building windows-amd64…"
+Write-Output "Building windows-amd64"
 Get-Date -Format "HH:mm:ss"
 $env:GOOS="windows"
 $env:GOARCH="amd64"
 go build -ldflags "$LDFLAGS" -o ./bin/ddc.exe
 Compress-Archive -Path ./bin/ddc.exe, ./bin/ddc.yaml -DestinationPath ./bin/ddc-windows-amd64.zip
+
+Write-Output "Building windows-arm64"
+Get-Date -Format "HH:mm:ss"
+$env:GOOS="windows"
+$env:GOARCH="arm64"
+go build -ldflags "$LDFLAGS" -o ./bin/ddc.exe
+Compress-Archive -Path ./bin/ddc.exe, ./bin/ddc.yaml -DestinationPath ./bin/ddc-windows-arm64.zip