+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## [0.6.1]
+
+### Added
+
+* binary for arm64 windows
+
+### Fixed
+
+* Windows collection and the ability to run tests on Windows
+
 ## [0.6.0]
 
 ### Added
@@ -310,6 +320,7 @@
 
 - able to capture logs, configuration and diagnostic data from dremio clusters deployed on Kubernetes and on-prem
 
+[0.6.1]: https://github.com/dremio/dremio-diagnostic-collector/compare/v0.6.0...v0.6.1
 [0.6.0]: https://github.com/dremio/dremio-diagnostic-collector/compare/v0.5.0...v0.6.0
 [0.5.0]: https://github.com/dremio/dremio-diagnostic-collector/compare/v0.4.0...v0.5.0
 [0.4.0]: https://github.com/dremio/dremio-diagnostic-collector/compare/v0.3.2...v0.4.0